refactor(commands): :recycle: Split subcommands module
Splited the chat_subcommands module into two new modules called subcommands_info and callbacks.
